% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{apotheke.df}
\alias{apotheke.df}
\title{Coordinates of 100 Random Pharmacies in Berlin}
\source{
© OpenStreetMap contributors - \url{https://www.openstreetmap.org/copyright/en}.
}
\description{
A data.frame of coordinates of 100 random pharmacies in Berlin. 
The projection is WGS 84.
}
