% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{apotheke.sp}
\alias{apotheke.sp}
\title{SpatialPointsDataFrame of 100 Random Pharmacies in Berlin}
\source{
© OpenStreetMap contributors - \url{https://www.openstreetmap.org/copyright/en}.
}
\description{
100 random pharmacies in Berlin. 
The projection is WGS 84 / UTM zone 34N.
}
