% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_osrm_profile.R
\name{find_osrm_profile}
\alias{find_osrm_profile}
\title{Locate an OSRM Lua profile (e.g. car.lua) in a host installation}
\usage{
find_osrm_profile(profile = "car.lua")
}
\arguments{
\item{profile}{A single string, the name of the Lua profile file
(e.g. \code{"car.lua"}). Defaults to \code{"car.lua"}.}
}
\value{
The normalized filesystem path to the profile.
}
\description{
By default OSRM ships profiles for "car", "bike" and "foot" in a
`profiles/` directory alongside the binaries.  This function will try to locate \code{osrm-routed} on the \code{PATH}, resolve symlinks, and look for sibling directories `share/osrm/profiles` and `share/osrm-backend/profiles`. IF that fails, it will try to fall back on \code{/usr/local/share/osrm/profiles},\code{/usr/local/share/osrm-backend/profiles}, \code{/usr/share/osrm/profiles}, and \code{/usr/share/osrm-backend/profiles}.
}
\examples{
\donttest{
find_osrm_profile("car.lua")
}

}
