% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm_stop_server.R
\name{osrm_stop_server}
\alias{osrm_stop_server}
\title{Stop an OSRM server started with `osrm_start_server`}
\usage{
osrm_stop_server(server)
}
\arguments{
\item{server}{A `processx::process` object returned by [osrm_start_server()].}
}
\value{
Invisibly returns the same `processx::process` object (now stopped).
}
\description{
Terminates an `osrm-routed` process that was launched by [osrm_start_server()].
}
