% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmIsodistance.R
\name{osrmIsometric}
\alias{osrmIsometric}
\title{deprecated}
\usage{
osrmIsometric(...)
}
\arguments{
\item{...}{deprecated}
}
\value{
deprecated
}
\description{
deprecated, use \link{osrmIsodistance} instead.
}
\keyword{internal}
