% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osrmTableErrors.R
\name{osrmTableErrors}
\alias{osrmTableErrors}
\title{Query A Distance Matrix and Extract Errors}
\usage{
osrmTableErrors(mat)
}
\arguments{
\item{mat}{A distance matrix.}
}
\value{
A list of two vectors of identifier: originErrors and destinationErrors.
}
\description{
Query a distance matrix and extracts errors in origins and destination.
}
\note{
An error is detected when at least half of distances are not found for an element.
}

