% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osrmTableOD.R
\name{osrmTableOD}
\alias{osrmTableOD}
\title{Query OSRM to Get Time Distances Between Origines and Destinations}
\usage{
osrmTableOD(dfo, ido, xo, yo, dfd, idd, xd, yd, limit = 100)
}
\arguments{
\item{dfo}{data frame containing origin points identifiers, longitudes and latitudes
(WGS84 projection).}

\item{ido}{identifier field in dfo.}

\item{xo}{longitude field in dfo.}

\item{yo}{latitude field in dfo.}

\item{dfd}{data frame containing destination points identifiers, longitudes and latitudes
(WGS84 projection).}

\item{idd}{identifier field in dfd.}

\item{xd}{longitude field in dfd.}

\item{yd}{latitude field in dfd.}

\item{limit}{table query limit of osrm API in use.}
}
\value{
A matrix of time distances (in minutes) between origins and
destinations is returned.
}
\description{
Query the OSRM API to get time distances marices between Origine
 and Destination points.
}

