% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmTripGeom.R
\name{osrmTripGeom}
\alias{osrmTripGeom}
\title{Get the travel geometry between multiple unordered points.}
\usage{
osrmTripGeom(userPoints, sp = FALSE)
}
\arguments{
\item{userPoints}{a SpatialPointsDataFrame of the waypoints, or a data.frame with points as rows
and at least 3 columns : id, lat and long (WGS84 decimal degrees)}

\item{sp}{if sp is TRUE the function returns a SpatialLinesDataFrame.}
}
\value{
A list with components

\describe{
\item{tripPoints}{A data.frame with the points lat/long (WGS84) and the step they're part of}
\item{geom}{Only when sp is TRUE, a SpatialLinesDataFrame (userPoints' CRS if there's one, WGS84 else),
containing a line for each step of the trip.}
\item{summary}{A list with 4 components : startingPoint, endingPoint, time (in minutes)
and distance (in kilometers)}
}
}
\description{
Build and send an OSRM API query to get the shortest travel geometry between multiple points.
This function interface the \emph{trip} OSRM service.
}
\seealso{
\link{osrmViarouteGeom}
}

