% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{src}
\alias{src}
\title{SpatialPointsDataFrame of 10 Communes in France And Belgium}
\source{
UMS RIATE
}
\description{
10 communes in France & Belgium. The projection is RGF93 / Lambert-93.
}

