% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osrmTable.R
\docType{package}
\name{osrm}
\alias{osrm}
\alias{osrm-package}
\title{Shortest Paths and Travel Time from OpenStreetMap via an OSRM API}
\description{
Query an OSRM service to obtain travel time between points in minutes. \cr
The package relies on the usage of a running OSRM instance
(located here: http://0.0.0.0:5000/)
}

