% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmIsochrone.R
\name{osrmIsochrone}
\alias{osrmIsochrone}
\title{Get a SpatialPolygonsDataFrame of Isochrones}
\usage{
osrmIsochrone(loc, breaks = seq(from = 0, to = 60, length.out = 7))
}
\arguments{
\item{loc}{a numeric vector of latitude and longitude (WGS84) or a 
SpatialPointsDataFrame or a SpatialPolygonsDataFrame of the origine point.}

\item{breaks}{a numeric vector of isochrone values.}
}
\value{
A SpatialPolygonsDateFrame of isochrones is returned. 
The data frame of the outputed SpatialPolygonsDataFrame contains four fields: 
id (id of each polygon), min and max (minimum and maximum breaks of the polygon), 
center (central values of classes)
}
\description{
Based on osrmTable this function buid a SpatialPolygonsDataFrame 
of isochrones.
}
\note{
This function uses the rgeos package.
}

