% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmIsochrone.R
\name{osrmIsometric}
\alias{osrmIsometric}
\title{Get Polygons of Isodistances}
\usage{
osrmIsometric(
  loc,
  breaks = seq(from = 0, to = 10000, length.out = 4),
  exclude = NULL,
  res = 30,
  returnclass = "sp"
)
}
\arguments{
\item{loc}{a numeric vector of longitude and latitude (WGS84), an sf object, 
a SpatialPointsDataFrame or a SpatialPolygonsDataFrame of the origine point.}

\item{breaks}{a numeric vector of isometric values (in meters).}

\item{exclude}{pass an optional "exclude" request option to the OSRM API.}

\item{res}{number of points used to compute isochrones, one side of the square 
grid, the total number of points will be res*res.}

\item{returnclass}{class of the returned polygons. Either "sp" of "sf".}
}
\value{
A SpatialPolygonsDateFrame or an sf MULTIPOLYGON of isochrones is returned. 
The data frame of the output contains four fields: 
id (id of each polygon), min and max (minimum and maximum breaks of the polygon), 
center (central values of classes).
}
\description{
Based on \code{\link{osrmTable}}, this function buids polygons 
 of isochrones.
}
\examples{
library(sf)
}
\seealso{
\link{osrmTable}
}
