% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osrmTable.R
\name{osrmTable}
\alias{osrmTable}
\title{Query OSRM to Get a Travel Time Matrix}
\usage{
osrmTable(df, id, x, y)
}
\arguments{
\item{df}{data frame containing points identifiers, longitudes and latitudes
(WGS84 projection).}

\item{id}{identifier field in df.}

\item{x}{longitude field in df.}

\item{y}{latitude field in df.}
}
\value{
x A square matrix of time travel (in minutes) is returned.
}
\description{
Query the OSRM API to get a travel time matrix between
points. This function interface the OSRM `table` service.
}
\note{
The public OSRM API does not allow more than 100
locations in distance table query. If you use an other API, make sure that
more than 100 locations are allowed (--max-table-size arg)
-m [ --max-table-size ] arg (=100)  Max. locations supported in distance table query
}

