% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osrmTableOD.R
\name{osrmTableOD}
\alias{osrmTableOD}
\title{Query OSRM to Get a Travel Time Matrix Between Origins and Destination}
\usage{
osrmTableOD(dfo, ido, xo, yo, dfd, idd, xd, yd, limit = 100)
}
\arguments{
\item{dfo}{data frame containing origin points identifiers, longitudes and latitudes
(WGS84 projection).}

\item{ido}{identifier field in dfo.}

\item{xo}{longitude field in dfo.}

\item{yo}{latitude field in dfo.}

\item{dfd}{data frame containing destination points identifiers, longitudes and latitudes
(WGS84 projection).}

\item{idd}{identifier field in dfd.}

\item{xd}{longitude field in dfd.}

\item{yd}{latitude field in dfd.}

\item{limit}{table query limit of osrm API in use. The default value (100) is
the public API limit.}
}
\value{
A matrix of time distances (in minutes) between origins and
destinations is returned.
}
\description{
Query the OSRM API to get travel time matrices between a set of
origin points and a set of destination points
}
\note{
This function is a wraper around osrmTable. Multiple calls allow to
obtain a matrix between different sets of origins and destinations.
}
\examples{
# Load data
data("com")
# Travel time matrix
distcom <- osrmTableOD(dfo = com[1:50,], ido = "comm_id", xo =  "lon", yo =  "lat",
                       dfd = com[51:100,], idd = "comm_id", xd = "lon", yd = "lat")
# First 5 rows and columns
distcom[1:5,1:5]
}

