% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osrmViaroute.R
\name{osrmViaroute}
\alias{osrmViaroute}
\title{Query OSRM to Get Travel Time and Travel Distance Between Two Points}
\usage{
osrmViaroute(xo, yo, xd, yd)
}
\arguments{
\item{xo}{longitude of the origine point.}

\item{yo}{latitude of the origine point.}

\item{xd}{longitude of the destination point.}

\item{yd}{latitude of the destination point.}
}
\value{
A named numeric vector is return. It contains travel time (in minutes)
and travel distance (in kilometers).
}
\description{
Query OSRM to get travel time and travel distance between two points.
This function interface the \emph{viaroute} OSRM service.
}

