% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osrmViaroute.R
\name{osrmViaroute}
\alias{osrmViaroute}
\title{Query OSRM to Get a Travel Distance, Time and Geometry}
\usage{
osrmViaroute(xo, yo, xd, yd)
}
\arguments{
\item{xo}{longitude of the origine point.}

\item{yo}{latitude of the origine point.}

\item{xd}{longitude of the destination point.}

\item{yd}{latitude of the destination point.}
}
\value{
a list is returned
}
\description{
Query the OSRM API to get a travel time matrix between
points. This function interface the OSRM `viaroute` service.
}

