% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osrmViarouteGeom.R
\name{osrmViarouteGeom}
\alias{osrmViarouteGeom}
\title{Query OSRM to Get the Travel Geometry Between Two Points}
\usage{
osrmViarouteGeom(xo, yo, xd, yd)
}
\arguments{
\item{xo}{longitude of the origine point.}

\item{yo}{latitude of the origine point.}

\item{xd}{longitude of the destination point.}

\item{yd}{latitude of the destination point.}
}
\value{
A data frame is return. It contains the longitudes and latitudes of
the travel path between the two points.
}
\description{
Query OSRM to get the travel geometry between two points.
This function interface the \emph{viaroute} OSRM service.
}

